requirements.txt
discord.py
main.py
openai>=1.3.9
python-dotenv
flask